+++ src/classes/Runner.cs
@@ -9,48 +9,87 @@ using Constants;
 /// </summary>
 public static class Runner
 {
+  /// <summary>
+  /// A class that contains properties for preparing and running programs from a list.
+  /// </summary>
+  public class PrepareProps
+  {
+    /// <summary>
+    /// The file path of the JSON file that contains the list of programs to run.
+    /// </summary>
+    public string FilePath { get; }
+
+    /// <summary>
+    /// The name of the list to run.
+    /// </summary>
+    public string ListName { get; }
+
+    /// <summary>
+    /// A value indicating whether to run the programs with elevated privileges.
+    /// </summary>
+    public bool GlobalElevated { get; }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="PrepareProps"/> class with the specified properties.
+    /// </summary>
+    /// <param name="FilePath">The file path of the JSON file that contains the list of programs to run.</param>
+    /// <param name="ListName">The name of the list to run.</param>
+    /// <param name="GlobalElevated">A value indicating whether to run the programs with elevated privileges.</param>
+    public PrepareProps(string? FilePath, string? ListName, bool? GlobalElevated)
+    {
+      this.FilePath = FilePath ?? string.Empty;
+      this.ListName = ListName ?? string.Empty;
+      this.GlobalElevated = GlobalElevated ?? false;
+    }
+  }
+
 
   /// <summary>
   /// Prepares and runs programs from a list.
   /// </summary>
-  /// <param name="filePath">The file path of the JSON file that contains the list of programs to run.</param>
-  /// <param name="listName">The name of the list to run</param>
-  public static void Prepare(string filePath, string listName)
+  /// <param name="props">The properties for preparing and running the programs.</param>
+  public static void Prepare(PrepareProps props)
   {
-    List<ListData>? data = JsonHelper.Load(filePath);
+    List<ListData>? data = JsonHelper.Load(props.FilePath);
 
     // If the list is empty or not found, return.
     if (data == null || data.Count() <= 0)
       return;
 
+    // Start the status display.
     AnsiConsole.Status()
       .Start("Starting...", ctx =>
       {
         ctx.Spinner(Spinner.Known.Dqpb);
         ctx.SpinnerStyle(Style.Parse("green"));
 
-        ctx.Status(Messages.RunningFromList(listName, filePath));
+        ctx.Status(Messages.RunningFromList(props.ListName, props.FilePath));
 
         // Find the list data for the specified list name.
-        ListData? listData = FindList(listName, data);
+        ListData? listData = FindList(props.ListName, data);
 
         // If the list data is not found, log an error and return.
         if (listData == null)
         {
-          Log.Error(Messages.ListNotFound(listName, filePath));
+          Log.Error(Messages.ListNotFound(props.ListName, props.FilePath));
           return;
         }
 
         // Run each program in the list data.
         foreach (ListProgram program in listData.Programs)
         {
+          bool elevated = false;
+
+          if (props.GlobalElevated || program.Elevated == true)
+            elevated = true;
+
           ctx.Status(Messages.TryingToRun(program.Name, program.Args));
 
           // Try to run the program using the specified path and arguments.
-          if (!RunProgram(program.Path, program.Args))
+          if (!RunProgram(program.Path, program.Args, elevated))
             // If the program cannot be run using the specified path and arguments,
             // try to run it using the command line
-            if (!RunCommand(program.Path, program.Args))
+            if (!RunCommand(program.Path, program.Args, elevated))
             {
               // If the program still cannot be run, log an error and continue
               Log.Error(Messages.RunningFailed(program.Name));
@@ -66,8 +105,9 @@ public static class Runner
   /// </summary>
   /// <param name="command">The command to run.</param>
   /// <param name="arguments">The arguments to pass to the command.</param>
+  /// <param name="elevated">Specifies whether to run the command with elevated privileges (e.g., as an administrator on Windows).</param>
   /// <returns>True if the command is successfully launched, otherwise false.</returns>
-  public static bool RunCommand(string command, string arguments)
+  public static bool RunCommand(string command, string arguments, bool elevated)
   {
     Process process = new Process();
     ProcessStartInfo startInfo = new ProcessStartInfo();
@@ -75,17 +115,24 @@ public static class Runner
     // Set the start info for the command line based on the operating system.
     if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
     {
+      // On Windows, use the "cmd" command with the "/c" argument to run the command.
       startInfo.FileName = "cmd";
       startInfo.Arguments = $"/c {command} {arguments}";
+
+      // If elevated privileges are requested, set the "Verb" property of the start info to "runas".
+      if (elevated)
+        startInfo.Verb = "runas";
     }
     else
     {
-      startInfo.FileName = "/bin/bash";
+      // On macOS and Linux, use "/bin/bash" to run the command,
+      // or "sudo" if elevated privileges are requested.
+      startInfo.FileName = elevated ? "sudo" : "/bin/bash";
       startInfo.Arguments = $"-c \"{command}\" {arguments}";
     }
 
     // Configure the start info for the process.
-    startInfo.UseShellExecute = false;
+    startInfo.UseShellExecute = elevated ? true : false;
     startInfo.CreateNoWindow = true;
     process.StartInfo = startInfo;
 
@@ -107,8 +154,9 @@ public static class Runner
   /// </summary>
   /// <param name="programPath">The path of the program to run.</param>
   /// <param name="arguments">The arguments to pass to the program.</param>
+  /// <param name="elevated">Whether to run the program with elevated privileges.</param>
   /// <returns>True if the program was successfully started, false otherwise.</returns>
-  public static bool RunProgram(string programPath, string arguments)
+  public static bool RunProgram(string programPath, string arguments, bool elevated)
   {
     string? program = programPath;
 
@@ -124,6 +172,16 @@ public static class Runner
 
     ProcessStartInfo startInfo = new ProcessStartInfo(program);
     startInfo.Arguments = arguments;
+    startInfo.UseShellExecute = elevated ? true : false;
+
+    // If elevated is true, configure the start info to run the program with elevated privileges.
+    if (elevated)
+      // If running on Windows, use the "runas" verb to elevate the process.
+      if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+        startInfo.Verb = "runas";
+      // If running on Linux or macOS, use sudo to elevate the process.
+      else
+        startInfo.FileName = "sudo";
 
     try
     {