+++ src/helpers/TextFile.cs
@@ -47,13 +47,17 @@ namespace TextFile
     /// </summary>
     public string Args { get; set; }
 
+    public bool? Elevated { get; set; }
+
     /// <summary>
     /// Initializes a new instance of the <see cref="ListProgram"/> class.
     /// </summary>
     public ListProgram()
     {
+      Name = string.Empty;
       Path = string.Empty;
       Args = string.Empty;
+      Elevated = false;
     }
   }
 
@@ -96,11 +100,25 @@ namespace TextFile
 
           foreach (JsonElement programElement in element.GetProperty("Programs").EnumerateArray())
           {
+            // Define a function to check if the "Elevated" property exists in the current JSON element.
+            Func<bool> IsElevated = () =>
+            {
+              try
+              {
+                return programElement.GetProperty("Elevated").GetBoolean();
+              }
+              catch (KeyNotFoundException)
+              {
+                return false;
+              }
+            };
+
             ListProgram program = new ListProgram
             {
-              Name = programElement.GetProperty("Name").GetString(),
+              Name = programElement.GetProperty("Name").GetString() ?? string.Empty,
               Path = programElement.GetProperty("Path").GetString() ?? string.Empty,
               Args = programElement.GetProperty("Args").GetString() ?? string.Empty,
+              Elevated = IsElevated()
             };
             listData.Programs.Add(program);
           }